+++ CHANGELOG.rst
@@ -10,11 +10,19 @@ Note that the issue IDs here refer to ones in the private CUBI GitLab.
 Unreleased
 ==========
 
+Added
+-----
+
+- **Appalerts**
+    - Display project badge in alert (#790)
+
 Changed
 -------
 
 - **General**
     - Unify app settings label notation (#793)
+- **Appalerts**
+    - Improve alert list layout (#790)
 - **Projectroles**
     - Improve login button locating in ``login_and_redirect_with_ui()`` (#796)
 
@@ -23,7 +31,6 @@ Fixed
 
 - **Appalerts**
     - Redundant HTML anchor in Dismiss All button (#788)
-
 
 
 v0.10.4 (2021-08-19)